fix: parseFloat voor entity-ID's — voorkomt truncatie van float-IDs
_maakId() geeft floats terug (bijv. 1717672345678.4371). parseInt() op
een data-*-attribuut dat zo'n waarde bevat, knipt het decimaal af en
breekt daarna alle id == waarde vergelijkingen.

- recepten.js: open-fiche, verwijder-recept, editId in saveRecept → parseFloat
- home.js: open-act-detail, toggle-todo-preview → parseFloat
- weekplanner.js: kies-plan-k → parseFloat

Bugfix: modal-backdrop listeners sloten alle 4 modals tegelijk;
elke listener sluit nu enkel zijn eigen modal.

+++ home.js
@@ -292,8 +292,8 @@ document.addEventListener('click', function (e) {
     case 'navigeer': location.href = el.dataset.href; break;
     case 'sluit-act-detail': sluitActDetail(); break;
     case 'wissel-persoon': wisselPersoon(el.dataset.key); break;
-    case 'open-act-detail': openActDetail(parseInt(el.dataset.id) || el.dataset.id); break;
-    case 'toggle-todo-preview': toggleTodo(parseInt(el.dataset.id) || el.dataset.id); break;
+    case 'open-act-detail': openActDetail(parseFloat(el.dataset.id) || el.dataset.id); break;
+    case 'toggle-todo-preview': toggleTodo(parseFloat(el.dataset.id) || el.dataset.id); break;
     case 'ga-todos': location.href = 'todos.html'; break;
   }
 });
